chore(lint): fixed all linter errors
Mostly line too long errors. I am ambivalent about it. On the one hand,
it's good to stick to the 80 chars standard just because standards are good.
And of course, it forced me to make some variable names less verbose, which
improved overall readability.
On the other hand, I would just have liked to up the max line length
to something closer to todays displays - 100 to 120 usually is the
desired number. This can't be done yet per project though, see
https://github.com/danielbrodin/atom-project-manager/issues/194