fix(block-cache): more safety with setCursor()
Previously, expander type would access the cursor
freely, which could hurt the consistency of the cache.

Now this is not possible anymore thanks to a setter.

Thinking about it ... as the cache has no state except
for the blocks it traverses, there might be a better solution,
one that doesn't require traversal on each set.